% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishSize}
\alias{fishSize}
\title{Fish size measurements from the Pacific Northwest.}
\format{A single column data frame with 3474 rows: \describe{
  \item{FishSize}{Recorded sizes of individual fish} }}
\source{
\url{TBD}
}
\usage{
fishSize
}
\description{
Example data set of fish size measurements from the Pacific Northwest, used
for illustrating the flexibility of the metalog distribution
}
\keyword{datasets}
