% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_method.R
\name{metalog_update}
\alias{metalog_update}
\title{Update the metalog with new data}
\usage{
metalog_update(m, new_data)
}
\arguments{
\item{m}{metalog object created using \code{metalog()}}

\item{x}{additional data to used to update the distribution shape}
}
\value{
A metalog object with updated parameters
}
\description{
Update the metalog with new data
}
\examples{
# example data
myData <- c(10,11,12,18)

# Create an unbounded metalog object
myMetalog <- metalog(myData,
                     term_lower_bound = 6)

newData <- c(13,17)

myUpdatedMetalog <- metalog_update(myMetalog, newData)
plot(myUpdatedMetalog)
}
