% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{plot.rmetalog}
\alias{plot.rmetalog}
\title{Plot of the metalog object}
\usage{
\method{plot}{rmetalog}(x, ...)
}
\arguments{
\item{x}{metalog object created using \code{r_metalog()}}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
A summary plot of the CDF and PDF for each term
}
\description{
Plot of the metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                       bounds=c(0, 60),
                       boundedness = 'b',
                       term_limit = 13)

plot(myMetalog)
}
