% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{r_metalog_plot}
\alias{r_metalog_plot}
\title{Plot of the metalog object}
\usage{
r_metalog_plot(my_metalog)
}
\arguments{
\item{my_metalog}{metalog object created using \code{r_metalog()}}
}
\value{
A summary plot of the CDF and PDF for each term
}
\description{
Plot of the metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 13)

r_metalog_plot(myMetalog)
}
