% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{r_metalog_sample}
\alias{r_metalog_sample}
\title{Create random samples from an rmetalog distribution object}
\usage{
r_metalog_sample(my_metalog, n = 1, term = 3)
}
\arguments{
\item{my_metalog}{metalog object created from \code{r_metalog()}}

\item{n}{number of observations (default is 1)}

\item{term}{which metalog distribution to sample from}
}
\value{
A numeric vector of n random samples from a selected distribution
}
\description{
Create random samples from an rmetalog distribution object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 9,
                      term_lower_bound = 9)

s <- r_metalog_sample(myMetalog, n=1000, term = 9)
hist(s)
}
