% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\docType{package}
\name{RMetalog}
\alias{RMetalog}
\alias{RMetalog-package}
\title{RMetalog: R implementation of the metalog distribution}
\description{
The \code{RMetalog} package implements the metalog distribution in \code{R}
}
