% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{rmetalog_quantile}
\alias{rmetalog_quantile}
\title{Generate quantiles from a probability from a metalog object}
\usage{
rmetalog_quantile(x, y, term = 3)
}
\arguments{
\item{x}{metalog object created from \code{r_metalog()}}

\item{y}{y vector of probabilities}

\item{term}{which metalog distribution to sample from}
}
\value{
A numeric vector of quantiles corresponding to the y probability vector
}
\description{
Generate quantiles from a probability from a metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 13)

s<-rmetalog_quantile(myMetalog)
}
