% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{rmetalog_sample}
\alias{rmetalog_sample}
\title{Create random samples from a metalog object}
\usage{
rmetalog_sample(x, n = 1, term = 3)
}
\arguments{
\item{x}{metalog object created from \code{r_metalog()}}

\item{n}{number of observations (default is 1)}

\item{term}{which metalog distribution to sample from}
}
\value{
A summary of the object and what it contains
}
\description{
Create random samples from a metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 13)

rmetalog_summary(myMetalog)
}
