% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_method.R
\name{summary.metalog}
\alias{summary.metalog}
\title{Summary of the metalog object}
\usage{
\method{summary}{metalog}(object, ...)
}
\arguments{
\item{object}{metalog object created from \code{metalog()}}

\item{...}{ignored; included for S3 generic/method consistency}
}
\value{
A summary of the object
}
\description{
Summary of the metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- metalog(fishSize$FishSize,
                     bounds=c(0, 60),
                     boundedness = 'b',
                     term_limit = 13)

summary(myMetalog)
}
