% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog_class_methods.R
\name{summary.rmetalog}
\alias{summary.rmetalog}
\title{Summary of the metalog object}
\usage{
\method{summary}{rmetalog}(x)
}
\arguments{
\item{x}{metalog object created from \code{r_metalog()}}
}
\value{
A summary of the object and what it contains
}
\description{
Summary of the metalog object
}
\examples{
# Load example data
data("fishSize")

# Create a bounded metalog object
myMetalog <- r_metalog(fishSize$FishSize,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 13)

summary(myMetalog)
}
