% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishSize}
\alias{fishSize}
\title{Fish size measurements from the Pacific Northwest.}
\format{
A single column data frame with 3474 rows: \describe{
  \item{FishSize}{Recorded sizes of individual steelhead trout} }
}
\source{
\url{http://www.metalogdistributions.com/}
}
\usage{
fishSize
}
\description{
Example data set of fish size measurements (in weight by pounds) from the Pacific Northwest, used
for illustrating the flexibility of the metalog distribution. This data set
is bi-modal because the fish contain two different populations, one salt and
two salt runs. The two salt, fish that have gone back to the ocean twice, are typically
larger.
}
\keyword{datasets}
