% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmetalog.R
\name{rMetalog}
\alias{rMetalog}
\title{Fit the Metalog distribution}
\usage{
rMetalog(x, step_len = 0.01, probs = 0, term_limit = 16, bounds = c(),
  boundedness = "u")
}
\arguments{
\item{x}{vector of numeric data}

\item{step_len}{size of steps to evaluate the distribution (between 0 and 1)}

\item{probs}{(Optional) probability quantiles, same length as \code{x}}

\item{term_limit}{integer between 5 and 30, specifying number of metalog
terms to evaluate (default: 16)}

\item{bounds}{numeric vector of size two, indicating lower/upper bounds}

\item{boundedness}{character string specifying unbounded, semi-bounded upper,
semi-bounded lower or bounded; accepts values \code{u}, \code{su},
\code{sl} and \code{b} (default: 'u')}
}
\value{
A list object with elements
\item{Y}{data frame}
\item{A}{data frame}
\item{M}{data frame}
\item{Validation}{character vector (?)}
}
\description{
Fit the Metalog distribution
}
\examples{
# Load example data
data("fishSize")

# Create a metalog object
myMetalog <- rMetalog(fishSize$FishSize,
                      step_len = .001,
                      bounds=c(0, 60),
                      boundedness = 'b',
                      term_limit = 16)
}
